requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

einops==0.7.0
timm==0.9.8
transformers==4.53.0
tokenizers==0.21.2

numpy==1.26.4
pandas==1.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.15.3

opencv-contrib-python==4.8.0.74
opencv-python==4.11.0.86
openslide-bin==4.0.0.8
openslide-python==1.4.2
pillow==11.2.1

tqdm==4.66.1
matplotlib==3.8.0
tabulate==0.9.0
h5py==3.14.0
